fix: typegoose sided mongodb errors
- split all models into client and server variants
