+++ src/lib/models/attendance.model.server.ts
@@ -1,0 +1,5 @@
+import { getModelForClass } from "@typegoose/typegoose";
+import { IAttendanceCode } from "./attendance.model";
+
+export const AttendanceCode = getModelForClass(IAttendanceCode);
+export type AttendanceCodeType = typeof AttendanceCode;
