+++ src/lib/models/nfc.model.server.ts
@@ -1,0 +1,5 @@
+import { getModelForClass } from "@typegoose/typegoose";
+import { INfcTag } from "./nfc.model";
+
+export const NfcTag = getModelForClass(INfcTag);
+export type NfcTagType = typeof NfcTag;
