+++ src/lib/models/user.model.server.ts
@@ -1,0 +1,5 @@
+import { getModelForClass } from "@typegoose/typegoose";
+import { IUser } from "./user.model";
+
+export const User = getModelForClass(IUser);
+export type UserType = typeof User;
